backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0 
sqlalchemy==2.0.23
pyodbc==5.01
python-dotenv==1.0.0 
pydantic==2.5.1
pydantic-settings==2.1.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
python-jose==3.3.0
aiofiles==23.2.1
requests==2.31.0
pytest==7.4.3
httpx==0.25.1
jinja2==3.1.2
websockets==12.0
uncertainties
uuid 
typing-extensions
fastapi-cache2==0.2.1
aioredis==2.0.1
email-validator==2.1.0
# requirements.txt additions
fastapi-mail==1.4.1  # For email functionality
python-jose[cryptography]  # For JWT handling
fastapi-cache2[redis]  # For caching
alembic==1.12.1  # For database migrations
JWT
zxcvbn
pyyaml
user-agents




